test(suites): classify CLI filesystem coverage
